requirements.txt
mesa>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
pm4py>=2.7.0
pytest>=7.0.0
ruff>=0.1.0
pyyaml>=6.0.0
matplotlib>=3.6.0
networkx>=3.0.0
